util: Move Wren functions in `instr` to `util`
