+++ src/classes/deps.c
@@ -14,7 +14,7 @@ WrenForeignMethodFn deps_bind_foreign_method(bool static_, char const* signature
 
 	// unknown
 
-	return unknown_foreign;
+	return wren_unknown_foreign;
 }
 
 // methods
