+++ src/classes/file.c
@@ -27,7 +27,7 @@ WrenForeignMethodFn file_bind_foreign_method(bool static_, char const* signature
 
 	// unknown
 
-	return unknown_foreign;
+	return wren_unknown_foreign;
 }
 
 // methods
