+++ src/classes/meta.c
@@ -24,7 +24,7 @@ WrenForeignMethodFn meta_bind_foreign_method(bool static_, char const* signature
 
 	// unknown
 
-	return unknown_foreign;
+	return wren_unknown_foreign;
 }
 
 // helpers
