+++ src/classes/rustc.h
@@ -176,5 +176,5 @@ static WrenForeignMethodFn rustc_bind_foreign_method(bool static_, char const* s
 
 	// unknown
 
-	return unknown_foreign;
+	return wren_unknown_foreign;
 }
